+++ src/main/java/chess/domain/board/BoardFactory.java
@@ -1,13 +1,13 @@
 package chess.domain.board;
 
-import chess.domain.piece.Bishop;
+import chess.domain.piece.linear.Bishop;
 import chess.domain.piece.Empty;
-import chess.domain.piece.King;
-import chess.domain.piece.Knight;
+import chess.domain.piece.immediate.King;
+import chess.domain.piece.immediate.Knight;
 import chess.domain.piece.Pawn;
 import chess.domain.piece.Piece;
-import chess.domain.piece.Queen;
-import chess.domain.piece.Rook;
+import chess.domain.piece.linear.Queen;
+import chess.domain.piece.linear.Rook;
 import chess.domain.piece.Side;
 import chess.domain.position.File;
 import chess.domain.position.Position;
