+++ src/main/java/chess/domain/piece/linear/Rook.java
@@ -1,16 +1,19 @@
-package chess.domain.piece;
+package chess.domain.piece.linear;
 
 import chess.domain.board.Board;
-import chess.domain.movepattern.QueenMovePattern;
+import chess.domain.movepattern.RookMovePattern;
+import chess.domain.piece.Piece;
+import chess.domain.piece.Side;
+import chess.domain.piece.Type;
 import chess.domain.position.Position;
 import java.util.List;
 
-public class Queen implements Piece {
+public class Rook implements Piece {
 
     private final LinearPiece linearPiece;
 
-    public Queen(final Side side) {
-        this.linearPiece = new LinearPiece(Type.QUEEN, side, List.of(QueenMovePattern.values()));
+    public Rook(final Side side) {
+        this.linearPiece = new LinearPiece(Type.ROOK, side, List.of(RookMovePattern.values()));
     }
 
     @Override
